Move Pretix API model classes to separate file
