+++ EuroPythonBot/helpers/pretix_connector.py
@@ -10,53 +10,13 @@ from pathlib import Path
 
 import aiofiles
 import aiohttp
-import pydantic
 from dotenv import load_dotenv
 
 from configuration import Config
 from error import AlreadyRegisteredError, NotFoundError
+from helpers.pretix_api_response_models import PretixItem, PretixItemVariation, PretixOrder
 
 _logger = logging.getLogger(f"bot.{__name__}")
-
-
-class PretixItem(pydantic.BaseModel):
-    """Item which can be ordered, e.g. 'Business', 'Personal', 'Education'."""
-
-    # https://docs.pretix.eu/en/latest/api/resources/items.html
-    id: int
-    names_by_locale: dict[str, str] = pydantic.Field(alias="name")
-    variations: list[PretixItemVariation]
-
-
-class PretixItemVariation(pydantic.BaseModel):
-    """Variation of item, e.g. 'Conference', 'Tutorial', 'Volunteer'."""
-
-    # https://docs.pretix.eu/en/latest/api/resources/item_variations.html
-    id: int
-    names_by_locale: dict[str, str] = pydantic.Field(alias="value")
-
-
-class PretixOrder(pydantic.BaseModel):
-    """Order containing one or more positions."""
-
-    # https://docs.pretix.eu/en/latest/api/resources/orders.html#order-resource
-    id: str = pydantic.Field(alias="code")
-    status: str
-    positions: list[PretixOrderPosition]
-
-    @property
-    def is_paid(self) -> bool:
-        # n: pending, p: paid, e: expired, c: canceled
-        return self.status == "p"
-
-
-class PretixOrderPosition(pydantic.BaseModel):
-    """Ordered position, e.g. a ticket or a T-shirt"""
-
-    # https://docs.pretix.eu/en/latest/api/resources/orders.html#order-position-resource
-    order_id: str = pydantic.Field(alias="order")
-    attendee_name: str | None
-    item_id: int = pydantic.Field(alias="item")
 
 
 def sanitize_username(username: str) -> str:
